pyproject.toml
[project]
name = "bookflow-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.128.0",
    "passlib[argon2]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.31",
]

[dependency-groups]
dev = ["pyrefly>=0.49.0", "ruff>=0.14.13", "taskipy>=1.14.1"]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']
ignore = ["PLR0917", "PLR0913"]

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pyrefly]
project-includes = ["src/**/*.py*", "src/**/*.ipynb", "main.py"]
project-excludes = ["tests/**/*", "venv/**/*", ".venv/**/*", "migrations/**/*"]
search-path = ["."]

[tool.pyrefly.modules]
fastapi = "typed"
pydantic = "typed"
sqlmodel = "typed"
uvicorn = "stubbed"

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
dev = 'fastapi dev src/main.py --reload-dir src'
